ingestion/pyproject.toml
[project]
name = "giggle-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "tenacity>=8.2.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.0",
    "tzdata>=2025.3",
    "googlemaps>=4.10.0",
    "pytz>=2025.2",
]
